+++ CHANGELOG.md
@@ -1,6 +1,8 @@
+# Your Minecraft, Your Way
+
 ## Changes:
-* Removed example sub-folder which caused initial runtime error on first install
+### Adds a toggle `settings.use-old-file-system`, It defaults to false however this is the system that was before the per-file change.
 
 ## Other:
-* [Feature Requests](https://github.com/Crazy-Crew/CrazyVouchers/issues)
+* [Feature Requests](https://github.com/Crazy-Crew/CrazyVouchers/discussions/categories/features)
 * [Bug Reports](https://github.com/Crazy-Crew/CrazyVouchers/issues)